refactor: add MyClassesStudent component and some changes in MyClassesView and StudentView for improved mobile experience
